Fix SSR SSG color mode flicker using next/cookies
